Fix isyntax bugs

See merge request empaia/integration/wsi-anon!37

+++ src/isyntax-io.h
@@ -9,7 +9,7 @@ static const char DOT_ISYNTAX[] = ".isyntax";
 // main functions
 struct metadata_attribute *get_attribute_isyntax(char *buffer, char *attribute);
 
-struct metadata *get_metadata_isyntax(file_handle *fp, int32_t header_size);
+struct metadata *get_metadata_isyntax(file_handle *fp, uint64_t header_size);
 
 struct wsi_data *get_wsi_data_isyntax(const char *filename);
